Add support for native bfloat16 on Arm
Eigen currently provides bfloat16 packet-math for NEON by emulating the
bfloat16 type using single precision floats. This patch adds support for
native bfloat16 on Arm plattforms when available (this should be the
case for armv8.3-a and newer).

When no native bfloat16 type is supported, the emulated version is still
used.

The packet-math part of this patch has been authored by Ashutosh Sharma
<ashutosh.sharma@amperecomputing.com>.