DRAFT: Rip out `make_coherent`, add CoherentCwiseBinaryOp.
The `make_coherent` approach really irks me, since it modifies `const`
inputs, and ends up requiring re-sizing expressions like
`CwiseNullaryOp` to get working.

Here we replace with an alternative binary op that checks the input
sizes, and if one has size zero then treats it as a zero input.

Marked as draft since there are likely performance considerations that
should be addressed.  This introduces an `if` at every coefficient
evaluation to check the input sizes.  There may be a better approach.
Also, for fixed-size inputs, we should probably use the original
`CwiseBinaryOp` to avoid this cost.